requirements.txt
Flask==2.3.2
boto3==1.28.0
python-dotenv==1.0.0
Werkzeug==2.3.6
Jinja2==3.1.2
click==8.1.3
itsdangerous==2.1.2
MarkupSafe==2.1.1
python-dateutil==2.8.2
